Rename: 더 명시적인 네이밍을 사용하도록 클래스명을 수정함
