--- src/main/java/com/growup/pms/auth/domain/RefreshToken.java
@@ -1,54 +1,0 @@
-package com.growup.pms.auth.domain;
-
-import com.growup.pms.common.BaseEntity;
-import com.growup.pms.user.domain.User;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.ForeignKey;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.SQLRestriction;
-
-@Entity
-@Getter
-@Table(name = "refresh_tokens")
-@SQLDelete(sql = "UPDATE refresh_tokens SET is_deleted = true WHERE id = ?")
-@SQLRestriction("is_deleted = false")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RefreshToken extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne
-    @JoinColumn(name = "user_id", nullable = false, updatable = false, foreignKey = @ForeignKey(name = "fk_refresh_token_user"))
-    private User user;
-
-    @Column(nullable = false, unique = true)
-    private String token;
-
-    @Column(nullable = false)
-    private LocalDateTime expiredAt;
-
-    @Builder
-    public RefreshToken(User user, String token, LocalDateTime expiryDate) {
-        this.user = user;
-        this.token = token;
-        this.expiredAt = expiryDate;
-    }
-
-    public void renewToken(String newToken, LocalDateTime newExpiryDate) {
-        this.token = newToken;
-        this.expiredAt = newExpiryDate;
-    }
-}
