+++ src/main/java/com/growup/pms/auth/service/mail/MailClient.java
@@ -1,7 +1,7 @@
 package com.growup.pms.auth.service.mail;
 
-import com.growup.pms.auth.service.dto.EmailDetails;
+import com.growup.pms.auth.service.dto.EmailSendCommand;
 
 public interface MailClient {
-    void sendEmail(EmailDetails emailDetails);
+    void sendEmail(EmailSendCommand emailDetails);
 }
