+++ src/main/java/com/growup/pms/common/aop/aspect/AuthorizationAspect.java
@@ -1,6 +1,6 @@
 package com.growup.pms.common.aop.aspect;
 
-import com.growup.pms.auth.domain.SecurityUser;
+import com.growup.pms.auth.controller.dto.SecurityUser;
 import com.growup.pms.common.aop.annotation.ProjectId;
 import com.growup.pms.common.aop.annotation.RequirePermission;
 import com.growup.pms.common.aop.annotation.TeamId;
